pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "docid"
version = "0.1.8"
description = "Deterministyczny generator identyfikatorów dokumentów z OCR"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Softreck", email = "info@softreck.dev"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
]
keywords = ["ocr", "document", "invoice", "ksef", "polish", "accounting"]
dependencies = [
    "pillow>=9.0.0",
    "pdf2image>=1.16.0",
    "numpy>=1.21.0",
    "pytesseract>=0.3.10",
    "opencv-python>=4.5.0",
    "reportlab>=3.6.0",
    "lxml>=4.6.0",
    "beautifulsoup4>=4.9.0",
    "pymupdf>=1.23.0",
    "paddleocr>=2.6.0",
    "paddlepaddle>=2.4.0",
]

[project.optional-dependencies]
paddle = [
    "paddleocr>=2.6.0",
    "paddlepaddle>=2.4.0",
]
tesseract = [
    "pytesseract>=0.3.10",
]
all = [
    "paddleocr>=2.6.0",
    "paddlepaddle>=2.4.0",
    "pytesseract>=0.3.10",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.scripts]
docid = "docid.cli:main"
docid-universal = "docid.cli_universal:main"

[project.urls]
Homepage = "https://github.com/softreck/docid"
Documentation = "https://github.com/softreck/docid#readme"
Repository = "https://github.com/softreck/docid"
Issues = "https://github.com/softreck/docid/issues"

[tool.setuptools.packages.find]
include = ["docid*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
